TESTES DE CONSUMO E REFATORAMENTO DE CODIGO

+++ code/PSNR.m
@@ -7,9 +7,9 @@ function  [dB] = PSNR(image1,image2, H,W,nFrames)
     MSE  = size(nFrames,1);
 
     for i = 1:1:nFrames
-        MSE(i)  = (sum(sum((image1(:,:,i) - image2(:,:,i)).^2))) / (H*W);
+        MSE(i)  = (sum(sum((double(image1(:,:,i)) - double(image2(:,:,i))).^2))) / (H*W);
 
     end
 
-    PSNR = 10*log10((255^2)./MSE);
+    PSNR = 10*log10((255.0^2)./MSE);
     dB = mean(PSNR);
